+++ db/migrations/20240625191529545_add_spent_at_index.sql
@@ -1,0 +1,7 @@
+-- +micrate Up
+-- SQL in section 'Up' is executed when this migration is applied
+CREATE INDEX IF NOT EXISTS entries_spent_at_idx ON entries (spent_at);
+
+-- +micrate Down
+-- SQL section 'Down' is executed when this migration is rolled back
+DROP INDEX IF EXISTS entries_spent_at_idx;